更改kSampleSize型別

+++ src/xamp_stream/include/stream/avfilestream.h
@@ -38,7 +38,7 @@ public:
 
 	std::string_view GetDescription() const noexcept override;
 
-	uint8_t GetSampleSize() const noexcept override;
+	uint32_t GetSampleSize() const noexcept override;
 
 	bool IsActive() const noexcept override;
